pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = "lab 6"
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
Flask-JWT = "^0.3.2"
Flask-Login = "^0.5.0"
Flask-SQLAlchemy = "^2.4"
Flask-WTF = "^0.14.3"
flask = "==1.0.2"
python = "^3.8"
email-validator = "^1.1.2"
WTForms = "^3.0.1"
